+++ src/color/naive_color_representation.h
@@ -1,7 +1,7 @@
 /// naive_color_representation.h
 
 /// realize color representation by the naive method
-/// using int to represent {0, 1, 2, 3}
+/// using size_t to represent {0, 1, 2, 3}
 
 #pragma once
 
@@ -12,15 +12,15 @@ namespace PlanarGraphColoring {
 
 class NaiveColorRepresentation : public ColorRepresentation {
 protected:
-  std::unordered_map<size_t, int> index2color_;
+  std::unordered_map<size_t, size_t> index2color_;
 public:
   NaiveColorRepresentation();
-  NaiveColorRepresentation(const std::vector<int>& colors);
+  NaiveColorRepresentation(const std::vector<size_t>& colors);
 
   virtual size_t size() const override;
-  virtual int get(const size_t index) const override; 
-  virtual std::vector<int> getVector() const override;
-  virtual void set(const size_t index, const int color) override;
+  virtual size_t get(const size_t index) const override; 
+  virtual std::vector<size_t> getVector() const override;
+  virtual void set(const size_t index, const size_t color) override;
   virtual void reset(const size_t index) override;
   virtual void show() const override;
   virtual bool valid() const override;
